Started getting yaw to work, height still needs fixing

+++ inc/Heli/heli/pid.h
@@ -11,9 +11,9 @@
 //*****************************************************************************
 typedef struct 
 {
-    uint16_t Kp; // Proportional Gain
-    uint16_t Ki; // Integral Gain
-    uint16_t Kd; // Derivative Gain
+    float Kp; // Proportional Gain
+    float Ki; // Integral Gain
+    float Kd; // Derivative Gain
 
     uint16_t max_Kp;
     uint16_t max_Ki;
@@ -30,7 +30,7 @@ typedef struct
 // Initalize PID controller
 //
 //*****************************************************************************
-controller_t* init_PID(uint16_t Kp, uint16_t Ki, uint16_t Kd, uint16_t max_Kp, uint16_t max_Ki, uint16_t max_Kd);
+controller_t* init_PID(float Kp, float Ki, float Kd, uint16_t max_Kp, uint16_t max_Ki, uint16_t max_Kd);
 
 
 //*****************************************************************************
@@ -38,6 +38,6 @@ controller_t* init_PID(uint16_t Kp, uint16_t Ki, uint16_t Kd, uint16_t max_Kp, u
 // Updates the PID controller
 //
 //*****************************************************************************
-uint16_t update_PID(controller_t* pid, int32_t error, uint16_t dT);
+uint16_t update_PID(controller_t* pid, int32_t error, float dT);
 
 #endif
